backend/pyproject.toml
[project]
name = "nnow"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.111",
    "uvicorn[standard]>=0.29",
    "python-multipart>=0.0.9",
    "httpx>=0.27",
    "pydantic>=2",
    "pydantic-settings>=2",
    "google-cloud-firestore>=2.16",
    "PyJWT>=2.8",
    "cryptography>=42",
]

[dependency-groups]
dev = [
    "mypy>=1.20.0",
    "ruff>=0.15.9",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "types-authlib>=1.6.9.20260402",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["auth", "shared", "todo"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
